feat(room): move room editor states and functions to useEditor hook
